pyproject.toml
[tool.poetry]
name = "genai-dataops-agent"
version = "0.1.0"
package-mode = true  # Change from false to true
readme = "README.md"
packages = [
    { include = "agent", from = "src" }
]


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.0.3"
numpy = "^1.24.3"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pyyaml = "^6.0.0"
jupyter = "^1.0.0"
ipykernel = "^6.0.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pandasai = "^3.0.0"
pandasai-litellm = "^0.0.1"
python-dotenv = "^1.0.0"
pandasai-sql = "^0.1.7"
slack-bolt = "^1.27.0"
slack-sdk = "^3.39.0"
structlog = "^25.5.0"
tenacity = "^9.1.2"

tiktoken = { version = "^0.7.0", optional = true }
psutil   = { version = "^6.0.0", optional = true }
pydantic-settings = "^2.12.0"

[tool.poetry.extras]
performance = ["tiktoken", "psutil"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"